refactor: split up query module
