+++ cognite/pygen/_core/generators.py
@@ -22,7 +22,7 @@ from cognite.client.data_classes.data_modeling.data_types import Enum
 from jinja2 import Environment, PackageLoader, select_autoescape
 from pydantic.version import VERSION as PYDANTIC_VERSION
 
-from cognite.pygen._query.extract_query_code import get_classes_code
+from cognite.pygen._query.extract_code import get_classes_code
 from cognite.pygen._version import __version__
 from cognite.pygen._warnings import PydanticNamespaceCollisionWarning
 from cognite.pygen.config import PygenConfig
