+++ cognite/pygen/_query/__init__.py
@@ -1,3 +1,3 @@
-from .query_execution import QueryExecutor
+from .interface import QueryExecutor
 
 __all__ = ["QueryExecutor"]
